locationous/requirements.txt
Django==5.1.2
psycopg2==2.9.10
python-ulid==3.0.0
celery==5.4.0
djangorestframework==3.15.2
redis==5.2.0
daphne==4.1.2